refactor(Roles.java): 역할에 대한 서비스 추가
딜러와 플레이어들에 대한 서비스를 수행하는 서비스 추가
